+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX05Z72KX4G20T57NDRXV00N"
+        "01KX0BE9WGY8R0FJA5GB9M0T4E"
     )
     r"""The unique identifier of the data source"""