+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KX05Z6Z0P4S3V4GZ36E58AF8"
+    id: Optional[str] = "01KX0BE9RQZWS25B30V330ZC78"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KX05Z6YA1EBS6SPZ9PAAQBW0"
+        "tool_01KX0BE9RF53JRJH3T5TPHVM6F"
     )
 
     display_name: Optional[str] = None
@@ -1488,7 +1488,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KX05Z6Y915BXGJE3PDJ84352"
+    id: Optional[str] = "01KX0BE9RE0VRWWSVCM07G54VJ"
 
     description: Optional[str] = None
 
@@ -1632,7 +1632,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KX05Z6Y87SQ4XM6XK2DCZ59V"
+        "tool_01KX0BE9RC4JHENYYZEYR20M9H"
     )
 
     display_name: Optional[str] = None
@@ -1932,7 +1932,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KX05Z6Y57RE8JFHNVN0EW5V8"
+        "tool_01KX0BE9RAT1YCBY9NAM94EQF8"
     )
 
     display_name: Optional[str] = None
@@ -2126,7 +2126,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KX05Z6Y4TNBSRB8D3N0N2JFJ"
+        "tool_01KX0BE9R89A8PZE5FT243ASW2"
     )
 
     display_name: Optional[str] = None
@@ -2324,7 +2324,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KX05Z6Y2Y2EE473W4X3FT91V"
+        "tool_01KX0BE9R6DCHBJJZ3X6ZPSFJV"
     )
 
     display_name: Optional[str] = None